requirements.txt
Werkzeug
Flask
numpy
gevent
h5py
tensorflow==2.8.2
gunicorn
Pillow
matplotlib
joblib
scikit-learn